requirements.txt
# Dependencias principales
openai>=1.30.0
anthropic>=0.28.0
google-genai>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
structlog>=24.2.0
tenacity>=8.3.0
rich>=13.7.0
aiosqlite>=0.20.0

# Dependencias de desarrollo y tests
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
